Creating Shader variations

+++ Engine/Common/PrimitiveTypes.h
@@ -18,6 +18,6 @@ using s8 = int8_t;
 using f32 = float;
 
 constexpr u64 u64_invalid_id{ 0xffff'ffff'ffff'ffff };
-constexpr u64 u32_invalid_id{ 0xffff'ffff };
-constexpr u64 u16_invalid_id{ 0xffff };
-constexpr u64 u8_invalid_id{ 0xff };
+constexpr u32 u32_invalid_id{ 0xffff'ffff };
+constexpr u16 u16_invalid_id{ 0xffff };
+constexpr u8 u8_invalid_id{ 0xff };
